Keep websocket diagnostics browser safe
